+++ services/client/app/(tabs)/groups/search.tsx
@@ -1,198 +1,40 @@
+import React, { useState } from "react";
 import {
-  Center,
-  RefreshControl,
-  ScrollView,
-  Spinner,
-  Text,
+  FormControl,
+  FormControlHelper,
+  FormControlHelperText,
+  FormControlLabel,
+  FormControlLabelText,
   VStack,
-  SafeAreaView,
+  Text,
   Button,
   ButtonText,
+  Center,
+  SafeAreaView,
+  ScrollView,
 } from "@gluestack-ui/themed";
-import React, { useCallback, useState } from "react";
-import GroupInformation from "../../../components/groups/GroupInformation";
-import { useMutation, useQuery } from "@apollo/client";
-import { usePrivateAuthContext } from "../../../hooks/auth";
-import { Stack, router, useFocusEffect } from "expo-router";
-import { gql } from "../../../__generated__";
+import LocationPicker from "../../../components/LocationPicker";
+import { Coordinates } from "../../../types/group";
+import { Stack, useRouter } from "expo-router";
 
-const GET_JOINABLE_GROUPS = gql(`
-  query GetNearestJoinableGroups(
-    $userStartLocation: CoordinatesInput!
-    $userEndLocation: CoordinatesInput!
-    $currentUserId: String!
-  ) {
-    nearestGroups(
-      userStartLocation: $userStartLocation
-      userEndLocation: $userEndLocation
-      where: {
-        and: [
-          { driver: { id: { neq: $currentUserId } } }
-          { passengers: { none: { id: { eq: $currentUserId } } } }
-        ]
-      }
-    ) {
-      id
-      startTime
-      days
-      startLocation {
-        latitude
-        longitude
-        distance(to: $userStartLocation)
-      }
-      endLocation {
-        latitude
-        longitude
-      }
-      totalSeats
-      passengers {
-        id
-      }
-    }
-  }
-`);
+const Search = () => {
+  const router = useRouter();
 
-const JOIN_GROUP = gql(`
-  mutation JoinGroup($groupId: Int!){
-    joinGroup(input: {id: $groupId}){
-      group {
-        id
-      }
-      errors {
-        ... on Error{
-          message
-        }
-      }
-    } 
-  }
-`);
-
-const Join = () => {
-  const { user } = usePrivateAuthContext();
-
-  // TODO: Should come dynamically from user input
-  const userStartLocation = {
+  // TODO: Get user location here
+  const [startLocation, setStartLocation] = useState<Coordinates>({
     latitude: 54.72090502968378,
     longitude: 25.28279660188754,
-  };
-
-  const userEndLocation = {
-    latitude: 54.70185784205424,
-    longitude: 25.259262733772623,
-  };
-
-  const {
-    loading: queryLoading,
-    error: queryError,
-    data: queryData,
-    refetch,
-  } = useQuery(GET_JOINABLE_GROUPS, {
-    variables: {
-      userStartLocation: userStartLocation,
-      userEndLocation: userEndLocation,
-      currentUserId: user.uid,
-    },
-    notifyOnNetworkStatusChange: true,
   });
 
-  const [
-    mutateFunction,
-    {
-      data: mutationData,
-      loading: mutationLoading,
-      error: mutationError,
-      called: mutationCalled,
-    },
-  ] = useMutation(JOIN_GROUP, {
-    onCompleted: (data) => {
-      if (data.joinGroup.errors) return;
-
-      router.navigate("/groups");
-    },
-  }); // Don't need to refetch GET_JOINABLE_GROUPS since we'll be navigated off this page
-
-  useFocusEffect(
-    useCallback(() => {
-      refetch();
-    }, [])
-  );
-
-  const [refreshing, setRefreshing] = useState<boolean>(false);
-
-  const onRefresh = useCallback(async () => {
-    setRefreshing(true);
-    await refetch();
-    setRefreshing(false);
-  }, []);
-
-  const body =
-    queryLoading || mutationLoading ? (
-      <Spinner />
-    ) : queryError || queryData === undefined ? (
-      <Text>Whoops! Ran into an error :/</Text>
-    ) : mutationCalled && (mutationError || mutationData === undefined) ? (
-      <Text>Whoops! Ran into an error when joining a group :/</Text>
-    ) : mutationData?.joinGroup.errors ? (
-      <>
-        <Text textAlign="center">Ran into an error!</Text>
-        {mutationData.joinGroup.errors.map((x) => (
-          <Text key={x.message} textAlign="center">
-            {x.message}
-          </Text>
-        ))}
-      </>
-    ) : queryData.nearestGroups.length === 0 ? (
-      <Text color="$secondary400" textAlign="center">
-        Seems like there are no new groups you can join :/
-      </Text>
-    ) : (
-      <VStack space="md">
-        {queryData.nearestGroups.map((group) => (
-          <GroupInformation
-            key={group.id}
-            group={{
-              startTime: group.startTime
-                .replace(/[PTM]/g, "")
-                .split("H")
-                .map((x) => x.padStart(2, "0"))
-                .join(":"),
-              days: group.days,
-              startLocation: group.startLocation,
-              endLocation: group.endLocation,
-              distanceFrom: group.startLocation.distance,
-              seats: {
-                total: group.totalSeats,
-                occupied: group.passengers.length,
-              },
-            }}
-            button={
-              <Button
-                variant="outline"
-                action="positive"
-                onPress={async () => {
-                  await mutateFunction({
-                    variables: {
-                      groupId: group.id,
-                    },
-                  });
-                }}
-              >
-                <ButtonText>Join</ButtonText>
-              </Button>
-            }
-          />
-        ))}
-      </VStack>
-    );
+  const [endLocation, setEndLocation] = useState<Coordinates>({
+    latitude: 54.68550466692954,
+    longitude: 25.26002211532131,
+  });
 
   return (
     <SafeAreaView h="$full">
-      <Stack.Screen options={{ title: "Join a group" }} />
-      <ScrollView
-        refreshControl={
-          <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
-        }
-      >
+      <Stack.Screen options={{ title: "Search" }} />
+      <ScrollView>
         <Center h="$full" p="$5">
           <VStack
             space="md"
@@ -201,7 +43,55 @@ const Join = () => {
             $md-w={"60%"}
             $lg-w={550}
           >
-            {body}
+            <FormControl>
+              <FormControlLabel mb="$1">
+                <FormControlLabelText>Start location</FormControlLabelText>
+              </FormControlLabel>
+              <Text>
+                {startLocation.latitude}, {startLocation.longitude}
+              </Text>
+              <LocationPicker
+                startingCoordinates={startLocation}
+                onConfirm={(newLocation) => setStartLocation(newLocation)}
+              />
+              <FormControlHelper>
+                <FormControlHelperText>
+                  Select where you want to leave
+                </FormControlHelperText>
+              </FormControlHelper>
+            </FormControl>
+            <FormControl>
+              <FormControlLabel mb="$1">
+                <FormControlLabelText>End location</FormControlLabelText>
+              </FormControlLabel>
+              <Text>
+                {endLocation.latitude}, {endLocation.longitude}
+              </Text>
+              <LocationPicker
+                startingCoordinates={endLocation}
+                onConfirm={(newLocation) => setEndLocation(newLocation)}
+              />
+              <FormControlHelper>
+                <FormControlHelperText>
+                  Select where you want to end up
+                </FormControlHelperText>
+              </FormControlHelper>
+            </FormControl>
+            <Button
+              onPress={() => {
+                router.push({
+                  pathname: "groups/result",
+                  params: {
+                    startLatitude: startLocation.latitude,
+                    startLongitude: startLocation.longitude,
+                    endLatitude: endLocation.latitude,
+                    endLongitude: endLocation.longitude,
+                  },
+                });
+              }}
+            >
+              <ButtonText>Search</ButtonText>
+            </Button>
           </VStack>
         </Center>
       </ScrollView>
@@ -209,4 +99,4 @@ const Join = () => {
   );
 };
 
-export default Join;
+export default Search;
